Big change, bringing everything back in to root and removing websocket completely
