+++ .env.sample
@@ -1,10 +1,10 @@
 APP_NAME='Stream Pet'
 NODE_ENV=local
-DEBUG=true
-LOG_LEVEL=verbose
+
+LOGIT_LEVEL=compressed
 
 OAUTH_TOKEN=
 BOT_USERNAME=
 CHANNEL_NAME=
 
-WEBSOCKET_PORT=1337
+PORT=1337
