--- server/lib/htmlEntities.js
@@ -1,8 +1,0 @@
-// Helper Function for escaping inputs
-var htmlEntities = function(str) {
-    return String(str)
-        .replace(/&/g, '&amp;').replace(/</g, '&lt;')
-        .replace(/>/g, '&gt;').replace(/"/g, '&quot;');
-}
-
-module.exports(htmlEntities);
